requirements.txt
fastapi
uvicorn
crewai
langchain-google-genai
langchain-groq
google-generativeai
groq
python-dotenv
pydantic
requests
streamlit
